web_gateway: carve dashboard_presence.rs (pure move)
The active-presence registry and the dashboard-control presence
connect/disconnect/make-active/cleanup lifecycle move to
web_gateway/dashboard_presence.rs. ActivePresence's connection_id and
direct_tx fields widen to pub(crate) for the listener's accesses.
